pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "circuithelper"
version = "0.1.0"
description = "Advanced circuit management plugin for NetBox"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "Digital Vortex LLC", email = "info@digitalvortex.com"}
]
keywords = ["netbox", "plugin", "circuit", "management", "ipam"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Telecommunications Industry",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Networking",
]

dependencies = [
    "netbox>=4.5.0",
    "fastkml>=1.0.0",
    "lxml>=4.9.0",
    "shapely>=2.0.0",
    "folium>=0.15.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "coverage>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/DigitalVortexLLC/circuithelper"
Documentation = "https://github.com/DigitalVortexLLC/circuithelper/blob/main/README.md"
Repository = "https://github.com/DigitalVortexLLC/circuithelper"
Issues = "https://github.com/DigitalVortexLLC/circuithelper/issues"
Changelog = "https://github.com/DigitalVortexLLC/circuithelper/releases"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "netbox.settings"
python_files = "test_*.py"
testpaths = ["tests"]

[tool.coverage.run]
source = ["circuithelper"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
